refactor(plugins): isolate pending conversation binding requests
